+++ doc/source/changes.rst
@@ -15,6 +15,7 @@ Versions
 
    * Added support for cross origin requests on ROA resources registry (/roa/resources).
    * Added support for exception serialized responses CORS compatible requests.
+   * Fix @RequiredScopes attribute for MVC controllers (now it automatically validates the security context before invoking underline method).
 
 * v0.6.0 (`Provide feedback <https://docs.google.com/forms/d/1tKBopU3lfDB_w8F4h7Rl1Rn4uydAJq-nha09L_ptJck/edit?usp=sharing>`_)