+++ doc/source/changes.rst
@@ -18,6 +18,7 @@ Versions
    * Fix @RequiredScopes attribute for MVC controllers (now it automatically validates the security context before invoking underline method).
    * Add support for headers appended to each response (support for global CORS configurable headers).
    * Improve implicit grant handler so that it supports non redirect responses.
+   * Fix cors for ROA fetch individual item.
 
 * v0.6.0 (`Provide feedback <https://docs.google.com/forms/d/1tKBopU3lfDB_w8F4h7Rl1Rn4uydAJq-nha09L_ptJck/edit?usp=sharing>`_)