+++ doc/source/changes.rst
@@ -14,11 +14,12 @@ Versions
 * v0.7.0 (Work in progress)
 
    * Added support for cross origin requests on ROA resources registry (/roa/resources).
-   * Added support for exception serialized responses CORS compatible requests.
-   * Fix @RequiredScopes attribute for MVC controllers (now it automatically validates the security context before invoking underline method).
+   * Added support for exception serialized responses CORS compatible requests.   
+   * Add CORS decorator for enabling CORS headers on custom controller routes.
    * Add support for headers appended to each response (support for global CORS configurable headers).
-   * Improve implicit grant handler so that it supports non redirect responses.
    * Fix cors for ROA fetch individual item.
+   * Fix @RequiredScopes attribute for MVC controllers (now it automatically validates the security context before invoking underline method).   
+   * Improve implicit grant handler so that it supports non redirect responses.
 
 * v0.6.0 (`Provide feedback <https://docs.google.com/forms/d/1tKBopU3lfDB_w8F4h7Rl1Rn4uydAJq-nha09L_ptJck/edit?usp=sharing>`_)